requirements.txt
asgiref==3.8.1
django>=3.2,<5.0 
djangorestframework==3.15.2
environ==1.0
paho-mqtt==2.1.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.1
    
    
